feat(action-bar): Add overflowActionsDisabled property.
