+++ tsconfig.json
@@ -1,5 +1,11 @@
 {
   "extends": "./tsconfig-base",
   "include": ["src", ".storybook/declarations.d.ts"],
-  "exclude": ["node_modules", "src/demos/**/*", "src/tests/**/*", "src/**/shared-list-tests.ts", "src/**/*.stories.ts"]
+  "exclude": [
+    "node_modules/**/*",
+    "src/demos/**/*",
+    "src/tests/**/*",
+    "src/**/shared-list-tests.ts",
+    "src/**/*.stories.ts"
+  ]
 }
